style(winds): satisfy ChildGuard clippy

+++ src/t060_fault_tests.rs
@@ -192,9 +192,8 @@ impl ChildGuard {
 
 impl Drop for ChildGuard {
     fn drop(&mut self) {
-        match self.child.try_wait() {
-            Ok(Some(_)) => return,
-            Ok(None) | Err(_) => {}
+        if let Ok(Some(_)) = self.child.try_wait() {
+            return;
         }
         let _ = self.child.kill();
         let _ = self.child.wait();
